Components: Extract TableOfContents as reusable component
