+++ editor/edit-post/header/header-toolbar/style.scss
@@ -52,3 +52,15 @@
 		}
 	}
 }
+
+.editor-header-toolbar .table-of-contents {
+	display: none;
+
+	.is-sidebar-opened & {
+		right: $sidebar-width + 16px;
+	}
+
+	@include break-small() {
+		display: block;
+	}
+}
